Remove manual padding of bistream

+++ crates/bevy_pbr/src/meshlet/asset.rs
@@ -39,7 +39,7 @@ pub const MESHLET_MESH_ASSET_VERSION: u64 = 1;
 #[derive(Asset, TypePath, Clone)]
 pub struct MeshletMesh {
     /// Quantized and bitstream-packed vertex positions for meshlet vertices.
-    pub(crate) vertex_positions: Arc<[u8]>,
+    pub(crate) vertex_positions: Arc<[u32]>,
     /// Octahedral-encoded and 2x16snorm packed normals for meshlet vertices.
     pub(crate) vertex_normals: Arc<[u32]>,
     /// Uncompressed vertex texture coordinates for meshlet vertices.
